+++ pl_bolts/models/self_supervised/moco/moco2_module.py
@@ -11,11 +11,10 @@ import torch.nn.functional as F
 import torchvision
 from torch import nn
 
-from pl_bolts.datamodules import CIFAR10DataModule, STL10DataModule
-from pl_bolts.datamodules.ssl_imagenet_datamodule import SSLImagenetDataModule
+from pl_bolts.datamodules import get_datamodule
 from pl_bolts.metrics import precision_at_k, mean
-from pl_bolts.models.self_supervised.moco.transforms import \
-    Moco2ImagenetTransforms, Moco2CIFAR10Transforms, Moco2STL10Transforms
+from pl_bolts.models.self_supervised.moco.transforms import (
+    Moco2ImagenetTransforms, Moco2CIFAR10Transforms, Moco2STL10Transforms)
 
 
 class MocoV2(pl.LightningModule):
@@ -230,17 +229,8 @@ class MocoV2(pl.LightningModule):
         return logits, labels
 
     def get_dataset(self, name):
-        if name == 'cifar10':
-            dataloaders = CIFAR10DataModule(self.hparams.data_dir, num_workers=self.hparams.num_workers)
-        elif name == 'stl10':
-            dataloaders = STL10DataModule(self.hparams.data_dir, num_workers=self.hparams.num_workers)
-        elif name == 'imagenet2012':
-            dataloaders = SSLImagenetDataModule(self.hparams.data_dir, num_workers=self.hparams.num_workers)
-        else:
-            raise FileNotFoundError(f'the {name} dataset is not supported. Subclass \'get_dataset to provide'
-                                    f'your own \'')
-
-        return dataloaders
+        extra = dict(meta_root=self.hparams.meta_root) if name == 'imagenet2012' else {}
+        return get_datamodule(name, data_dir=self.hparams.data_dir, num_workers=self.hparams.num_workers, **extra)
 
     def prepare_data(self):
         self.dataset.prepare_data()
